master-control/master-backend/requirements.txt
fastapi==0.104.1
aiofiles==23.2.1
uvicorn==0.24.0
python-socketio==5.11.1
websockets==12.0
python-dotenv==1.0.0
httpx==0.27.0
python-multipart==0.0.9